Degrib: fix int overflow added in today commit
git-svn-id: https://svn.osgeo.org/gdal/trunk@40775 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/grib/degrib/degrib/clock.c
@@ -2460,7 +2460,7 @@ double Clock_AddMonthYear (double refTime, int incrMonth, int incrYear)
    double d_remain;
    int i;
 
-   if( !(fabs(refTime) < SEC_DAY * 365 * 10000) )
+   if( !(fabs(refTime) < (double)SEC_DAY * 365 * 10000) )
        return 0;
 
    totDay = (sInt4) floor (refTime / SEC_DAY);
